requirements.txt
scikit-learn
apache-airflow==2.8.1
apache-airflow-providers-mysql==5.5.0
pymysql==1.1.0
sqlalchemy==1.4.52
pandas==2.1.4
numpy==1.26.3
pyarrow==14.0.2
requests==2.31.0
great_expectations==1.16.0
matplotlib==3.8.2
seaborn==0.13.1
jupyter==1.0.0
python-dotenv==1.0.0
